move dnd to prop row
