+++ packages/client/src/pages/MainPage/containers/PropGroup/PropGroupRow/PropGroupRow.tsx
@@ -1,6 +1,7 @@
 import { IEntity, IProp } from "@shared/types";
 import { AttributeIcon, Button } from "components";
-import React, { useEffect, useState } from "react";
+import React, { useEffect, useRef, useState } from "react";
+import { useDrop, DropTargetMonitor, XYCoord, useDrag } from "react-dnd";
 import {
   FaCaretDown,
   FaCaretUp,
@@ -8,13 +9,16 @@ import {
   FaTrashAlt,
   FaUnlink,
 } from "react-icons/fa";
-import { useAppSelector } from "redux/hooks";
+import { setDraggedPropRow } from "redux/features/propGroup/draggedPropRowSlice";
+import { useAppDispatch, useAppSelector } from "redux/hooks";
 import { excludedSuggesterEntities } from "Theme/constants";
 import {
   AttributeGroupDataObject,
   classesPropType,
   classesPropValue,
   DraggedPropRowItem,
+  DragItem,
+  ItemTypes,
 } from "types";
 import { EntitySuggester, EntityTag } from "../..";
 import { AttributesGroupEditor } from "../../AttributesEditor/AttributesGroupEditor";
@@ -45,6 +49,10 @@ interface IPropGroupRow {
   openDetailOnCreate: boolean;
 
   parentId: string;
+  id?: string;
+  index?: number;
+  moveProp: (dragIndex: number, hoverIndex: number) => void;
+  itemType?: ItemTypes;
 }
 
 export const PropGroupRow: React.FC<IPropGroupRow> = ({
@@ -59,10 +67,14 @@ export const PropGroupRow: React.FC<IPropGroupRow> = ({
   addProp,
   movePropDown,
   movePropUp,
+  moveProp,
   userCanEdit,
   territoryActants = [],
   openDetailOnCreate = false,
   parentId,
+  id,
+  index,
+  itemType,
 }) => {
   const propTypeEntity: IEntity = entities[prop.type.id];
   const propValueEntity = entities[prop.value.id];
@@ -81,237 +93,312 @@ export const PropGroupRow: React.FC<IPropGroupRow> = ({
     }
   }, [draggePropRow]);
 
-  return (
-    <StyledGrid key={level + "|" + order} tempDisabled={tempDisabled}>
-      <StyledPropLineColumn level={level} isTag={propTypeEntity ? true : false}>
-        <StyledFaGripVertical />
-        {propTypeEntity ? (
-          <EntityTag
-            actant={propTypeEntity}
-            fullWidth
-            button={
+  const dispatch = useAppDispatch();
+  const ref = useRef<HTMLDivElement>(null);
+
+  const [{ handlerId }, drop] = useDrop({
+    accept: itemType ? itemType : ItemTypes.PROP_ROW,
+    collect(monitor) {
+      return {
+        handlerId: monitor.getHandlerId(),
+      };
+    },
+    hover(item: DragItem, monitor: DropTargetMonitor) {
+      if (!ref.current) {
+        return;
+      }
+      const dragIndex: number = item.index;
+      const hoverIndex: number | undefined = index;
+
+      if (dragIndex === hoverIndex) {
+        return;
+      }
+
+      const hoverBoundingRect = ref.current?.getBoundingClientRect();
+      const hoverMiddleY =
+        (hoverBoundingRect.bottom - hoverBoundingRect.top) / 2;
+      const clientOffset = monitor.getClientOffset();
+      const hoverClientY = (clientOffset as XYCoord).y - hoverBoundingRect.top;
+      if (hoverIndex === undefined) {
+        return;
+      }
+      if (dragIndex < hoverIndex && hoverClientY < hoverMiddleY) {
+        return;
+      }
+
+      if (dragIndex > hoverIndex && hoverClientY > hoverMiddleY) {
+        return;
+      }
+      moveProp(dragIndex, hoverIndex);
+      item.index = hoverIndex;
+    },
+  });
+
+  const [{ isDragging }, drag] = useDrag({
+    item: { type: itemType ? itemType : ItemTypes.PROP_ROW, id, index },
+    collect: (monitor) => ({
+      isDragging: monitor.isDragging(),
+    }),
+  });
+
+  drag(drop(ref));
+
+  useEffect(() => {
+    if (isDragging) {
+      dispatch(setDraggedPropRow({ id, parentId, index }));
+    } else {
+      dispatch(setDraggedPropRow({}));
+    }
+  }, [isDragging]);
+
+  const renderPropRow = () => {
+    return (
+      <StyledGrid key={level + "|" + order} tempDisabled={tempDisabled}>
+        <StyledPropLineColumn
+          level={level}
+          isTag={propTypeEntity ? true : false}
+        >
+          <StyledFaGripVertical />
+          {propTypeEntity ? (
+            <EntityTag
+              actant={propTypeEntity}
+              fullWidth
+              button={
+                <Button
+                  key="d"
+                  icon={<FaUnlink />}
+                  color="plain"
+                  inverted={true}
+                  tooltip="unlink actant"
+                  onClick={() => {
+                    updateProp(prop.id, {
+                      type: {
+                        ...prop.type,
+                        ...{ id: "" },
+                      },
+                    });
+                  }}
+                />
+              }
+            />
+          ) : (
+            <EntitySuggester
+              territoryActants={territoryActants}
+              onSelected={(newSelectedId: string) => {
+                updateProp(prop.id, {
+                  type: {
+                    ...prop.type,
+                    ...{ id: newSelectedId },
+                  },
+                });
+              }}
+              openDetailOnCreate={openDetailOnCreate}
+              categoryTypes={classesPropType}
+              inputWidth={90}
+              excludedEntities={excludedSuggesterEntities}
+            />
+          )}
+          <StyledPropButtonGroup>
+            {prop.type.logic == "2" ? (
               <Button
-                key="d"
-                icon={<FaUnlink />}
-                color="plain"
+                key="neg"
+                tooltip="Negative logic"
+                color="success"
                 inverted={true}
-                tooltip="unlink actant"
-                onClick={() => {
-                  updateProp(prop.id, {
-                    type: {
-                      ...prop.type,
-                      ...{ id: "" },
-                    },
-                  });
-                }}
+                noBorder
+                icon={<AttributeIcon attributeName={"negation"} />}
               />
-            }
-          />
-        ) : (
-          <EntitySuggester
-            territoryActants={territoryActants}
-            onSelected={(newSelectedId: string) => {
-              updateProp(prop.id, {
+            ) : (
+              <div />
+            )}
+          </StyledPropButtonGroup>
+        </StyledPropLineColumn>
+        <StyledPropLineColumn isTag={propValueEntity ? true : false}>
+          {propValueEntity ? (
+            <EntityTag
+              actant={propValueEntity}
+              fullWidth
+              button={
+                <Button
+                  key="d"
+                  icon={<FaUnlink />}
+                  tooltip="unlink actant"
+                  color="plain"
+                  inverted={true}
+                  onClick={() => {
+                    updateProp(prop.id, {
+                      value: {
+                        ...prop.value,
+                        ...{ id: "" },
+                      },
+                    });
+                  }}
+                />
+              }
+            />
+          ) : (
+            <EntitySuggester
+              territoryActants={[]}
+              onSelected={(newSelectedId: string) => {
+                updateProp(prop.id, {
+                  value: {
+                    ...prop.type,
+                    ...{ id: newSelectedId },
+                  },
+                });
+              }}
+              openDetailOnCreate={openDetailOnCreate}
+              categoryTypes={classesPropValue}
+              inputWidth={90}
+              excludedEntities={excludedSuggesterEntities}
+            />
+          )}
+          <StyledPropButtonGroup>
+            {prop.value.logic == "2" ? (
+              <Button
+                key="neg"
+                tooltip="Negative logic"
+                color="success"
+                inverted={true}
+                noBorder
+                icon={<AttributeIcon attributeName={"negation"} />}
+              />
+            ) : (
+              <div />
+            )}
+          </StyledPropButtonGroup>
+        </StyledPropLineColumn>
+
+        <StyledPropLineColumn>
+          <StyledPropButtonGroup>
+            <AttributesGroupEditor
+              modalTitle={`Property attributes`}
+              disabledAllAttributes={!userCanEdit}
+              propTypeActant={propTypeEntity}
+              propValueActant={propValueEntity}
+              excludedSuggesterEntities={excludedSuggesterEntities}
+              classesPropType={classesPropType}
+              classesPropValue={classesPropValue}
+              updateProp={updateProp}
+              statementId={prop.id}
+              data={{
+                statement: {
+                  elvl: prop.elvl,
+                  certainty: prop.certainty,
+                  logic: prop.logic,
+                  mood: prop.mood,
+                  moodvariant: prop.moodvariant,
+                  bundleOperator: prop.bundleOperator,
+                  bundleStart: prop.bundleStart,
+                  bundleEnd: prop.bundleEnd,
+                },
                 type: {
-                  ...prop.type,
-                  ...{ id: newSelectedId },
+                  elvl: prop.type.elvl,
+                  logic: prop.type.logic,
+                  virtuality: prop.type.virtuality,
+                  partitivity: prop.type.partitivity,
                 },
-              });
-            }}
-            openDetailOnCreate={openDetailOnCreate}
-            categoryTypes={classesPropType}
-            inputWidth={90}
-            excludedEntities={excludedSuggesterEntities}
-          />
-        )}
-        <StyledPropButtonGroup>
-          {prop.type.logic == "2" ? (
-            <Button
-              key="neg"
-              tooltip="Negative logic"
-              color="success"
-              inverted={true}
-              noBorder
-              icon={<AttributeIcon attributeName={"negation"} />}
+                value: {
+                  elvl: prop.value.elvl,
+                  logic: prop.value.logic,
+                  virtuality: prop.value.virtuality,
+                  partitivity: prop.value.partitivity,
+                },
+              }}
+              handleUpdate={(newData: AttributeGroupDataObject) => {
+                const newDataObject = {
+                  ...newData.statement,
+                  ...newData,
+                };
+                const { statement, ...statementPropObject } = newDataObject;
+                updateProp(prop.id, statementPropObject);
+              }}
+              userCanEdit={userCanEdit}
             />
-          ) : (
-            <div />
-          )}
-        </StyledPropButtonGroup>
-      </StyledPropLineColumn>
-      <StyledPropLineColumn isTag={propValueEntity ? true : false}>
-        {propValueEntity ? (
-          <EntityTag
-            actant={propValueEntity}
-            fullWidth
-            button={
+
+            {(level === "1" || level === "2") && (
               <Button
-                key="d"
-                icon={<FaUnlink />}
-                tooltip="unlink actant"
+                key="add"
+                icon={<FaPlus />}
+                tooltip="add child prop"
                 color="plain"
                 inverted={true}
                 onClick={() => {
-                  updateProp(prop.id, {
-                    value: {
-                      ...prop.value,
-                      ...{ id: "" },
-                    },
-                  });
+                  addProp(prop.id);
                 }}
               />
-            }
-          />
-        ) : (
-          <EntitySuggester
-            territoryActants={[]}
-            onSelected={(newSelectedId: string) => {
-              updateProp(prop.id, {
-                value: {
-                  ...prop.type,
-                  ...{ id: newSelectedId },
-                },
-              });
-            }}
-            openDetailOnCreate={openDetailOnCreate}
-            categoryTypes={classesPropValue}
-            inputWidth={90}
-            excludedEntities={excludedSuggesterEntities}
-          />
-        )}
-        <StyledPropButtonGroup>
-          {prop.value.logic == "2" ? (
+            )}
             <Button
-              key="neg"
-              tooltip="Negative logic"
-              color="success"
-              inverted={true}
-              noBorder
-              icon={<AttributeIcon attributeName={"negation"} />}
-            />
-          ) : (
-            <div />
-          )}
-        </StyledPropButtonGroup>
-      </StyledPropLineColumn>
-
-      <StyledPropLineColumn>
-        <StyledPropButtonGroup>
-          <AttributesGroupEditor
-            modalTitle={`Property attributes`}
-            disabledAllAttributes={!userCanEdit}
-            propTypeActant={propTypeEntity}
-            propValueActant={propValueEntity}
-            excludedSuggesterEntities={excludedSuggesterEntities}
-            classesPropType={classesPropType}
-            classesPropValue={classesPropValue}
-            updateProp={updateProp}
-            statementId={prop.id}
-            data={{
-              statement: {
-                elvl: prop.elvl,
-                certainty: prop.certainty,
-                logic: prop.logic,
-                mood: prop.mood,
-                moodvariant: prop.moodvariant,
-                bundleOperator: prop.bundleOperator,
-                bundleStart: prop.bundleStart,
-                bundleEnd: prop.bundleEnd,
-              },
-              type: {
-                elvl: prop.type.elvl,
-                logic: prop.type.logic,
-                virtuality: prop.type.virtuality,
-                partitivity: prop.type.partitivity,
-              },
-              value: {
-                elvl: prop.value.elvl,
-                logic: prop.value.logic,
-                virtuality: prop.value.virtuality,
-                partitivity: prop.value.partitivity,
-              },
-            }}
-            handleUpdate={(newData: AttributeGroupDataObject) => {
-              const newDataObject = {
-                ...newData.statement,
-                ...newData,
-              };
-              const { statement, ...statementPropObject } = newDataObject;
-              updateProp(prop.id, statementPropObject);
-            }}
-            userCanEdit={userCanEdit}
-          />
-
-          {(level === "1" || level === "2") && (
-            <Button
-              key="add"
-              icon={<FaPlus />}
-              tooltip="add child prop"
+              key="delete"
+              icon={<FaTrashAlt />}
+              tooltip="remove prop row"
               color="plain"
               inverted={true}
               onClick={() => {
-                addProp(prop.id);
+                removeProp(prop.id);
               }}
             />
-          )}
-          <Button
-            key="delete"
-            icon={<FaTrashAlt />}
-            tooltip="remove prop row"
-            color="plain"
-            inverted={true}
-            onClick={() => {
-              removeProp(prop.id);
-            }}
-          />
-          <Button
-            key="up"
-            inverted
-            disabled={firstRowinGroup}
-            icon={<FaCaretUp />}
-            tooltip="move prop up"
-            color="plain"
-            onClick={() => {
-              movePropUp(prop.id);
-            }}
-          />
-          <Button
-            key="down"
-            inverted
-            disabled={lastRowinGroup}
-            icon={<FaCaretDown />}
-            tooltip="move prop down"
-            color="plain"
-            onClick={() => {
-              movePropDown(prop.id);
-            }}
-          />
-          {prop.logic == "2" ? (
             <Button
-              key="neg"
-              tooltip="Negative logic"
-              color="success"
-              inverted={true}
-              noBorder
-              icon={<AttributeIcon attributeName={"negation"} />}
+              key="up"
+              inverted
+              disabled={firstRowinGroup}
+              icon={<FaCaretUp />}
+              tooltip="move prop up"
+              color="plain"
+              onClick={() => {
+                movePropUp(prop.id);
+              }}
             />
-          ) : (
-            <div />
-          )}
-          {prop.bundleOperator ? (
             <Button
-              key="oper"
-              tooltip="Logical operator type"
-              color="success"
-              inverted={true}
-              noBorder
-              icon={prop.bundleOperator}
+              key="down"
+              inverted
+              disabled={lastRowinGroup}
+              icon={<FaCaretDown />}
+              tooltip="move prop down"
+              color="plain"
+              onClick={() => {
+                movePropDown(prop.id);
+              }}
             />
-          ) : (
-            <div />
-          )}
-        </StyledPropButtonGroup>
-      </StyledPropLineColumn>
-    </StyledGrid>
+            {prop.logic == "2" ? (
+              <Button
+                key="neg"
+                tooltip="Negative logic"
+                color="success"
+                inverted={true}
+                noBorder
+                icon={<AttributeIcon attributeName={"negation"} />}
+              />
+            ) : (
+              <div />
+            )}
+            {prop.bundleOperator ? (
+              <Button
+                key="oper"
+                tooltip="Logical operator type"
+                color="success"
+                inverted={true}
+                noBorder
+                icon={prop.bundleOperator}
+              />
+            ) : (
+              <div />
+            )}
+          </StyledPropButtonGroup>
+        </StyledPropLineColumn>
+      </StyledGrid>
+    );
+  };
+
+  return (
+    <>
+      {!tempDisabled ? (
+        <div ref={ref} data-handler-id={handlerId}>
+          {renderPropRow()}
+        </div>
+      ) : (
+        <>{renderPropRow()}</>
+      )}
+    </>
   );
 };
