--- packages/client/src/pages/MainPage/containers/PropGroup/PropGroupRowDndWrapper/PropGroupRowDndWrapperStyles.tsx
@@ -1,9 +1,0 @@
-import styled from "styled-components";
-
-interface StyledDndWrapper {
-  tempDisabled?: boolean;
-}
-export const StyledDndWrapper = styled.div<StyledDndWrapper>`
-  background-color: ${({ tempDisabled }) =>
-    tempDisabled ? "hotpink" : "transparent"};
-`;
