+++ packages/client/src/pages/MainPage/containers/PropGroup/SecondLevelPropGroup/SecondLevelPropGroup.tsx
@@ -1,15 +1,14 @@
 import { IProp } from "@shared/types";
 import update from "immutability-helper";
 import React, { useCallback, useEffect, useState } from "react";
-import { ItemTypes } from "types";
-import { PropGroupRowDndWrapper } from "../PropGroupRowDndWrapper/PropGroupRowDndWrapper";
 
 interface SecondLevelPropGroup {
   prop1: IProp;
   renderSecondLevelPropRow: (
     prop2: IProp,
     pi2: number,
-    prop1: IProp
+    prop1: IProp,
+    moveProp: (dragIndex: number, hoverIndex: number) => void
   ) => JSX.Element;
 }
 export const SecondLevelPropGroup: React.FC<SecondLevelPropGroup> = ({
@@ -35,18 +34,9 @@ export const SecondLevelPropGroup: React.FC<SecondLevelPropGroup> = ({
 
   return (
     <>
-      {secondLevelProps.map((prop2: IProp, pi2: number) => (
-        <PropGroupRowDndWrapper
-          key={prop1.id + "|" + pi2}
-          id={prop2.id}
-          lvl={2}
-          parentId={prop1.id}
-          index={pi2}
-          itemType={ItemTypes.PROP_ROW2}
-          moveProp={moveProp}
-          renderPropRow={renderSecondLevelPropRow(prop2, pi2, prop1)}
-        />
-      ))}
+      {secondLevelProps.map((prop2: IProp, pi2: number) =>
+        renderSecondLevelPropRow(prop2, pi2, prop1, moveProp)
+      )}
     </>
   );
 };
