+++ packages/client/src/types.ts
@@ -111,6 +111,7 @@ export enum ItemTypes {
   TAG = "TAG",
   STATEMENT_ROW = "STATEMENT_ROW",
   ACTANT_ROW = "ACTANT_ROW",
+  PROP_ROW = "PROP_ROW",
   PROP_ROW1 = "PROP_ROW1",
   PROP_ROW2 = "PROP_ROW2",
   PROP_ROW3 = "PROP_ROW3",
